# pawn AI 생성 성공
- projectile 생성 버그 수정
- pawn AI 생성 추적, 공격, 회피
- 시작시 쿨타임 버그 수정
- 테이블 수정

+++ Assets/Scripts/Data/Data.Contents.cs
@@ -302,15 +302,15 @@ namespace Data
         public string icon;
         public Define.ESkillType skillType;
         public Define.ETargetType targetType;
-        public int manaAmount;
-        public int minRange;
-        public int maxRange;
-        public int splashRange;
-        public int coolTime;
+        public float manaAmount;
+        public float minRange;
+        public float maxRange;
+        public float splashRange;
+        public float coolTime;
         public string effectStr;
         public Define.EPawnAniTriger aniTriger;
         public Define.EPawnAniState motionAni;
-        public int motionDuration;
+        public float motionDuration;
         public string projectile;
         public int[] arr_affect = new int[3];
     }
